Performance Entrada, transferencia de produtos, backend e frontend

+++ src/main/java/br/com/pi/projeto_RD/controller/FilialController.java
@@ -12,6 +12,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigInteger;
+
 @RestController
 public class FilialController {
 
@@ -33,7 +35,7 @@ public class FilialController {
     //LISTAR POR ID
     @GetMapping("/filial/{codigo}")
     @ApiOperation(value = "Listar filiais por ID")
-    public ResponseEntity buscarPorId(@PathVariable("codigo") Long codigo) {
+    public ResponseEntity buscarPorId(@PathVariable("codigo") BigInteger codigo) {
         return ResponseEntity.ok().body(service.buscarPorId(codigo));
     }
