+++ src/main/java/br/com/pi/projeto_RD/model/dto/TrocaMotivoDTO.java
@@ -3,10 +3,12 @@ package br.com.pi.projeto_RD.model.dto;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class TrocaMotivoDTO {
 
-    private Long idFilial;
+    private BigInteger idFilial;
     private String nmFilial;
     private double totalTroca;
     private double totalDesistencia;
