+++ src/main/java/br/com/pi/projeto_RD/repository/FilialRepository.java
@@ -3,5 +3,7 @@ package br.com.pi.projeto_RD.repository;
 import br.com.pi.projeto_RD.model.entity.FilialEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FilialRepository extends JpaRepository<FilialEntity, Long> {
+import java.math.BigInteger;
+
+public interface FilialRepository extends JpaRepository<FilialEntity, BigInteger> {
 }
